+++ packages/cosmwasm-stargate/src/signingcosmwasmclient.ts
@@ -527,10 +527,9 @@ export class SigningCosmWasmClient extends CosmWasmClient {
     const signedGasLimit = Int53.fromString(signed.fee.gas).toNumber();
     const signedSequence = Int53.fromString(signed.sequence).toNumber();
     const signedAuthInfoBytes = makeAuthInfoBytes(
-      [pubkey],
+      [{ pubkey, sequence: signedSequence }],
       signed.fee.amount,
       signedGasLimit,
-      signedSequence,
       signMode,
     );
     return TxRaw.fromPartial({
@@ -564,7 +563,7 @@ export class SigningCosmWasmClient extends CosmWasmClient {
     };
     const txBodyBytes = this.registry.encode(txBody);
     const gasLimit = Int53.fromString(fee.gas).toNumber();
-    const authInfoBytes = makeAuthInfoBytes([pubkey], fee.amount, gasLimit, sequence);
+    const authInfoBytes = makeAuthInfoBytes([{ pubkey, sequence }], fee.amount, gasLimit);
     const signDoc = makeSignDoc(txBodyBytes, authInfoBytes, chainId, accountNumber);
     const { signature, signed } = await this.signer.signDirect(signerAddress, signDoc);
     return TxRaw.fromPartial({